cmd/cli: move cobra commands to separated file
So each command initialization/logic can be read/update more easily.
